+++ src/hyperlight_guest_bin/src/guest_function/call.rs
@@ -23,6 +23,7 @@ use hyperlight_common::flatbuffer_wrappers::function_types::{
 };
 use hyperlight_common::flatbuffer_wrappers::guest_error::{ErrorCode, GuestError};
 use hyperlight_guest::error::{HyperlightGuestError, Result};
+use hyperlight_guest::transport::DispatchAction;
 use hyperlight_guest::{bail, transport};
 use tracing::instrument;
 
@@ -95,12 +96,16 @@ pub(crate) fn internal_dispatch_function() {
         tracing::span!(tracing::Level::INFO, "internal_dispatch_function").entered()
     };
 
-    let (cid, function_call) = transport::with_ctx(|ctx| ctx.recv_h2g_call())
-        .expect("Function call deserialization failed");
+    let dispatch = transport::with_ctx(|ctx| ctx.recv_h2g_dispatch())
+        .expect("H2G dispatch deserialization failed");
 
-    let result = call_guest_function(function_call)
-        .map_err(|error| GuestError::new(error.kind, error.message));
-    let result = FunctionCallResult::new(result);
+    let result = match dispatch {
+        DispatchAction::Call(cid, fc) => {
+            let res = call_guest_function(fc).map_err(|err| GuestError::new(err.kind, err.message));
+            Some((cid, FunctionCallResult::new(res)))
+        }
+        DispatchAction::SnapshotCheckpoint => None,
+    };
 
     // All this tracing logic shall be done right before the call to `hlt` which is done after this
     // function returns
@@ -119,6 +124,10 @@ pub(crate) fn internal_dispatch_function() {
         hyperlight_guest_tracing::flush();
     }
 
-    transport::with_ctx(|ctx| ctx.send_h2g_result(cid, result))
-        .expect("Failed to send function call result");
+    match result {
+        Some((cid, result)) => transport::with_ctx(|ctx| ctx.send_h2g_result(cid, result))
+            .expect("Failed to send function call result"),
+        None => transport::with_ctx(|ctx| ctx.prepare_snapshot())
+            .expect("Failed to prepare snapshot transport"),
+    }
 }